Working on layout
Been messing around with the layout. Need to improve the css styling and
relative positioning. Currently, each player box is using the same css
and 'hand'. May need to: add a new class to the same style as the hand,
make the player box cards smaller, rearrange size of boxes and their
borders.