Render cards from a JSON formatted object
Created a function render_game() that will delete all divs with the card
class and recreate them based on the new game_state (which is currently
inline in the javascript function). When the user owns the card (their
username matches the owner of the card), it goes to their hand. Only
when the card is on the table does the x,y,z position matter. Otherwise,
the card is not rendered because it is in someone else's hand. For the
x,y,z positions, changing the values to percentages will be trivial (add
a % in the css values).
Need to do/figure out:
-how to get the JSON file
-implement the owner boxes (display dynamically for each player, #
cards, change color when their turn)
-count the number of cards in each players hands
-prevent players from flipping/firing events when it is not their turn
-May have to implement someway to format cards in a deck (offset x
coordinate by 2px per card or something)