Display game name in Footer
The game name now displays in the center of the footer.  I wasn't sure
the exact moment the gameCache contains the game name because it would
be undefined for the most part. Currently in render_game which will
render the game name every time.

Discovered that the full game name is not being stored into
gameCache.name and that the hand is not being rendered in the hand on
refresh in a single player game. Not sure if the same bug applies to a
multiplayer game.